requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli==1.27.38
flake8
python-dotenv>=0.5.1
torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu116
numpy 
datetime
os
argparse
matplotlib